+++ mnozenie.c
@@ -1,5 +1,5 @@
-float mnozenie (float a,float b, float c){
-float wynik;
+double mnozenie (double a,double b, double c){
+double wynik;
 wynik=a*b*c;
 return wynik;
 }
